Move attributions into sources route
This is somewhat unfortunate because passing the link in as
a string means it’s not Ember-y, so clicking it causes a
reload rather than a smooth transition where the map remains.
Maybe I can figure that out but for now it’s better to have
the attribution section not be huge on mobile.