Move map bounds calculation to controller
This prevents a mysterious error when trying to revisit the
map route after going back to the index… but now another
similar error is happening regarding a closure property 😳

Probs making the map a component would help 😞

Error while processing route: map.index Assertion Failed: You attempted to update [object Object].closure to "Tue Sep 01 2020 00:00:00 GMT-0500", but it is being tracked by a tracking context, such as a template, computed property, or observer. In order to make sure the context updates properly, you must invalidate the property when updating it. You can mark the property as `@tracked`, or use `@ember/object#set` to do this.