Add guard against adding closures/cuts twice
The previous change was a red herring; this was the real
problem, that the collated closures and cuts were being
added again, while not being tracked. Tracking isn’t needed
since the data is static; really this would be better part
of preprocessing, but that doesn’t work with date objects.